# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.GetCategoryDTO:
    properties:
      category_id:
        type: integer
      category_name:
        type: string
      category_slug:
        type: string
    type: object
  dto.GetOrderDTO:
    properties:
      order_id:
        type: integer
      order_items:
        items:
          $ref: '#/definitions/dto.GetOrderItemDTO'
        type: array
      order_status:
        type: string
      order_total:
        type: number
      user_id:
        type: integer
    type: object
  dto.GetOrderItemDTO:
    properties:
      order_item_id:
        type: integer
      order_item_quantity:
        type: integer
      order_item_total:
        type: number
      product:
        $ref: '#/definitions/dto.GetProductDTO'
    type: object
  dto.GetProductDTO:
    properties:
      category:
        $ref: '#/definitions/dto.GetCategoryDTO'
      product_description:
        type: string
      product_id:
        type: integer
      product_image:
        type: string
      product_name:
        type: string
      product_price:
        type: number
      seller_id:
        type: integer
    type: object
...
  handler.ErrorResponse:
    properties:
      message:
        example: validation failed
        type: string
    type: object
...
  handler.OrderIDResponse:
    properties:
      order_id:
        example: 101
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: 'REST API for marketplace: products, categories, cart, orders, auth
    and roles.'
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Marketplace API
  version: "1.0"
paths:
...
  /orders:
    get:
      description: Get all orders of the authenticated user.
      produces:
      - application/json
      responses:
        "200":
          description: List of user orders
          schema:
            items:
              $ref: '#/definitions/dto.GetOrderDTO'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List user orders
      tags:
      - Orders
    post:
      description: 'Create an order from the current cart contents. Rate limit: 2
        req/min.'
      produces:
      - application/json
      responses:
        "200":
          description: Created order ID
          schema:
            $ref: '#/definitions/handler.OrderIDResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "429":
          description: Too many requests
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create order
      tags:
      - Orders
...
securityDefinitions:
  BearerAuth:
    description: 'JWT access token. Example: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"